requirements-development.txt
-e .
pytest>3.0
pytest-django
pytest-factoryboy
faker
django-filter
tox
tox-pyenv
tox-travis
mock
